added a parent class for rooms, enemies, and containers to contain all public methods and fields
