+++ DungeonExplorer.csproj
@@ -43,7 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="entities\TestEnemy.cs" />
+    <Compile Include="entities\Enemies\TestEnemy.cs" />
     <Compile Include="Main\Game.cs" />
     <Compile Include="entities\LivingEntity.cs" />
     <Compile Include="entities\Player.cs" />
@@ -52,6 +52,8 @@
     <Compile Include="Main\Test.cs" />
     <Compile Include="Map\Containers\ParentContainer.cs" />
     <Compile Include="Map\Containers\TestContainer.cs" />
+    <Compile Include="entities\Enemies\ParentEnemy.cs" />
+    <Compile Include="Map\Rooms\ParentRoom.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Map\Rooms\DefaultRoom.cs" />
   </ItemGroup>
