+++ Map/Containers/ParentContainer.cs
@@ -8,6 +8,8 @@ namespace DungeonExplorer
 {
 
     public class ParentContainer
+    // a class for all container types to inherit from
+    // all public functions and fields of an container should be implemented in here
     {
         public string Name { get; protected set; }
     }
