+++ entities/Enemies/ParentEnemy.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DungeonExplorer
+{
+    public class ParentEnemy : LivingEntity
+    // a class for all enemy types to inherit from
+    // all public functions and fields of an enemy should be implemented in here
+    {
+    }
+}
